# static/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API to interact with [Wakapi](https://wakapi.dev)\n\n## Authentication\nSet header `Authorization` to your API Key encoded as Base64 and prefixed with `Basic`\n**Example:** `Basic ODY2NDhkNzQtMTljNS00NTJiLWJhMDEtZmIzZWM3MGQ0YzJmCg==`",
        "title": "Wakapi API",
        "contact": {
            "name": "Ferdinand Mütsch",
            "url": "https://github.com/muety",
            "email": "ferdinand@muetsch.io"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://github.com/muety/wakapi/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/api/compat/wakatime/v1/users/{user}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mimics https://wakatime.com/developers#users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wakatime"
                ],
                "summary": "Retrieve the given user",
                "operationId": "get-wakatime-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to fetch (or 'current')",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.UserViewModel"
                        }
                    }
                }
            }
        },
...
        "/api/compat/wakatime/v1/users/{user}/stats/{range}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mimics https://wakatime.com/developers#stats",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wakatime"
                ],
                "summary": "Retrieve statistics for a given user",
                "operationId": "get-wakatimes-tats",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to fetch data for (or 'current')",
                        "name": "user",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "today",
                            "yesterday",
                            "week",
                            "month",
                            "year",
                            "7_days",
                            "last_7_days",
                            "30_days",
                            "last_30_days",
                            "12_months",
                            "last_12_months",
                            "any"
                        ],
                        "type": "string",
                        "description": "Range interval identifier",
                        "name": "range",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.StatsViewModel"
                        }
                    }
                }
            }
        },
        "/api/compat/wakatime/v1/users/{user}/summaries": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mimics https://wakatime.com/developers#summaries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wakatime"
                ],
                "summary": "Retrieve WakaTime-compatible summaries",
                "operationId": "get-wakatime-summaries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to fetch data for (or 'current')",
                        "name": "user",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "today",
                            "yesterday",
                            "week",
                            "month",
                            "year",
                            "7_days",
                            "last_7_days",
                            "30_days",
                            "last_30_days",
                            "12_months",
                            "last_12_months",
                            "any"
                        ],
                        "type": "string",
                        "description": "Range interval identifier",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date (e.g. '2021-02-07')",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date (e.g. '2021-02-08')",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.SummariesViewModel"
                        }
                    }
                }
            }
        },
...
        "/api/users/{user}/statusbar/today": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mimics https://wakatime.com/api/v1/users/current/statusbar/today. Have no official documentation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wakatime"
                ],
                "summary": "Retrieve summary for statusbar",
                "operationId": "statusbar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to fetch data for (or 'current')",
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.StatusBarViewModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "v1.StatsData": {
            "type": "object",
            "properties": {
                "daily_average": {
                    "type": "number"
                },
                "days_including_holidays": {
                    "type": "integer"
                },
                "editors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "end": {
                    "type": "string"
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "machines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "operating_systems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "start": {
                    "type": "string"
                },
                "total_seconds": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "v1.StatsViewModel": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.StatsData"
                }
            }
        },
        "v1.StatusBarViewModel": {
            "type": "object",
            "properties": {
                "cached_at": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/v1.SummariesData"
                }
            }
        },
        "v1.SummariesData": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "dependencies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "editors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "grand_total": {
                    "$ref": "#/definitions/v1.SummariesGrandTotal"
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "machines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "operating_systems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "projects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesEntry"
                    }
                },
                "range": {
                    "$ref": "#/definitions/v1.SummariesRange"
                }
            }
        },
        "v1.SummariesEntry": {
            "type": "object",
            "properties": {
                "digital": {
                    "type": "string"
                },
                "hours": {
                    "type": "integer"
                },
                "minutes": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "percent": {
                    "type": "number"
                },
                "seconds": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "total_seconds": {
                    "type": "number"
                }
            }
        },
        "v1.SummariesGrandTotal": {
            "type": "object",
            "properties": {
                "digital": {
                    "type": "string"
                },
                "hours": {
                    "type": "integer"
                },
                "minutes": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                },
                "total_seconds": {
                    "type": "number"
                }
            }
        },
        "v1.SummariesRange": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                }
            }
        },
        "v1.SummariesViewModel": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.SummariesData"
                    }
                },
                "end": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
        "v1.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_email_confirmed": {
                    "type": "boolean"
                },
                "is_email_public": {
                    "type": "boolean"
                },
                "last_heartbeat_at": {
                    "type": "string"
                },
                "last_plugin_name": {
                    "type": "string"
                },
                "last_project": {
                    "type": "string"
                },
                "modified_at": {
                    "type": "string"
                },
                "timezone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
        "v1.UserViewModel": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/v1.User"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}